rimphony: work on cleaning up distribution functions
The definition of the power law distribution function was way overcomplicated,
so I've simplified it.

But. There's a 1/(gamma gamma beta) term in the function that I thought I
could just ditch; but it affects the gamma derivative of f when evaluating
absorption coefficients. In principle one should be able to simplify it away,
but I've been having trouble and it might be a bit too late at night for me to
take derivatives with square roots.

Also, the tests are failing fairly frequently now. It seems that taking all of
the constants out of the integrands has gotten the numerics to a point where
they do not converge robustly. I'll deal with this; there are only a couple of
tuning parameters, and I think that problems should only be cropping up if
some things are erroneously depending on signs or absolute magnitudes.
Although based on how some computations are massively slower than others, I
think there are probably some improvements to be made in general.