Cargo.toml: hack to get binaries without AVX instructions
It turns out that the special-fun crate always compiles its C routines with
the `-march=native` compiler flag, which leads to code that uses AVX
instructions on the compilers that I am using. This means that my binaries
often die on Odyssey, where many nodes do not have AVX.

I solve this by using a forked version that has the relevant compiler flag
removed. It would be better to compile the C code with the `gcc` crate, but
that's more work than I want to put in right now. Issue filed as

https://github.com/vks/special-fun/issues/3