leung-bessel: improve a corner case for the Bessel function evaluator
I ran into some problematic integrals that boiled down to a problem in the
Leung Bessel evaluator. The clearest problem was that `Jn_prime` gave very
different answers for (n, z) = (1.037977420000000000e+11,
1.037977332932058411e+11) and the very close (n, z) =
(1.037977420000000000e+11, 1.037977332932060089e+11). (These have the same *n*
value but slightly different *z* values.) This problem in turn arose because
`Jn` had a discontinuity at these inputs.

Unsurprisingly, this discontinuity turned out to be because that particular
pair of inputs crosses the boundary where `my_Bessel_J` chooses which
approximation to use. It tries to switch over in the region where the two
approximations agree well, but (1) even a small discontinuity can lead to a
large derivative discontinuity and (2) I find that the expression deciding
where to make the transition is not as accurate as it could be.

I've updated the code to now make a smooth (linear) transition in a narrow
region, rather than a discontinuous transition at a particular point. I should
preserve the supporting work for posterity somewhere, but for now I want to
get this improved version deployed. Initial tests seem to show that it indeed
fixes the problematic cases I had run into.