src/symphony.rs: try to avoid the n >= 1e15 death point
As per the previous commit, if we hit n > 1e15, we have to give up because we
can't reliably compute a Bessel derivative. If we expand the bounds of our *n*
integration too much, GSL might evaluate the integrand for such a huge *n*
even though we don't actually need to. Therefore, if we're getting close to
such a big *n*, take steps to keep the integrations narrower in the hopes that
we can converge before crossing the death threshold.